requirements.txt
pandas>=2.0.0
numpy>=1.24.0
optuna>=3.5.0
openai-whisper>=20231117
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.28.0
streamlit>=1.32.0
pytest>=8.0.0
python-dotenv>=1.0.0
yt-dlp>=2024.3.10
Pillow>=10.0.0
scipy>=1.12.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
requests>=2.31.0
stable-baselines3>=2.3.0
gymnasium>=0.29.0
torch>=2.2.0
